testes

+++ flask/file_api/app.py
@@ -1,26 +1,36 @@
 from flask import Flask, request, jsonify
+import helper_function as hf
+import glob
+
 
 app = Flask(__name__)
 app.config['JSON_SORT_KEYS'] = False
 
-list_files = []
+# check if file.txt exists if not empty list
+list_files = glob.glob('*.txt') # list of files created
 
 @app.route("/") # home page
 def home():
-    return "Create File API"
+    return "Hello, This is the first Task: Create an API with Flask"
 
 # Endpoint to create a file
 @app.route("/create_file", methods=["POST"])
 def create_file():
     post_res = request.get_json()
     filename = post_res.get("file_name")
-    list_files.append(post_res)
+    
+    if hf.check_file_exists(filename):
+        return jsonify(
+            Mensagem= "Arquivo ja existe, por favor escolha outro nome",
+            Arquivos = list_files)
+
+    list_files.append(filename)
     with open(filename, "w") as f:
         f.write(f"O Arquivo criado foi {filename}")
         
     return jsonify(
         Mensagem= "Lista de Arquivos Criados",
-        Data = list_files)
+        Arquivos = list_files)
     
 # Endpoint to list all the files created   
 @app.route("/file_list", methods=["GET"])
@@ -35,6 +45,12 @@ def read_filelist():
 def read_file():
     post_res = request.get_json()
     filename = post_res.get("file_name")
+    
+    if hf.check_file_exists(filename) == False:
+        return jsonify(
+            Mensagem= "Arquivo nao existe, por favor escolha outro nome",
+            Arquivos = list_files)
+        
     with open(filename, "r") as f:
         contents = f.readlines()
         
@@ -48,6 +64,11 @@ def update_file():
     post_res = request.get_json()
     filename = post_res.get("file_name")
     message = post_res.get("message")
+    
+    if hf.check_file_exists(filename) == False:
+        return jsonify(
+            Mensagem= "Arquivo nao existe, por favor escolha outro nome",
+            Arquivos = list_files)
     
     with open(filename, "a") as f:
         f.write(f" {message}")
